pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "minsearch>=0.0.5",
    "openai>=1.108.2",
    "pydantic>=2.12.3",
    "pydantic-ai==1.0.9",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "arize-phoenix>=5.7.0",
    "openinference-semantic-conventions>=0.1.11",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "pandas>=2.2.0",
]
